fix: Fix build fix: Fix composer theme
