+++ packages/react-theme-default/src/components/SCInlineComposer.tsx
@@ -1,13 +1,18 @@
 const Component = {
   styleOverrides: {
     root: ({theme}: any) => ({
+      marginBottom: theme.spacing(2),
       '& .SCInlineComposer-content, & .SCInlineComposer-content:last-child': {
         padding: theme.spacing(0.5, 1),
+        display: 'flex',
+        flexDirection: 'row',
+        justifyContent: 'space-between',
         [theme.breakpoints.up('sm')]: {
           padding: theme.spacing(0.5, 1, 0.5, 2)
         },
         '& .SCInlineComposer-input': {
           display: 'flex',
+          flexGrow: 2,
           '& .MuiButton-root': {
             fontWeight: theme.typography.fontWeightLight,
             letterSpacing: '0.17px',
@@ -17,6 +22,10 @@ const Component = {
               background: 'white',
               border: 'none'
             }
+          },
+          ['& .MuiButton-text']: {
+            justifyContent: 'flex-start',
+            textTransform: 'none'
           }
         },
         '& .SCInlineComposer-actions': {
@@ -33,6 +42,10 @@ const Component = {
           alignItems: 'center',
           [theme.breakpoints.up('sm')]: {
             marginLeft: theme.spacing(2)
+          },
+          '& .MuiAvatar-root': {
+            width: theme.selfcommunity.user.avatar.sizeMedium,
+            height: theme.selfcommunity.user.avatar.sizeMedium
           }
         }
       }
