+++ packages/react-ui/src/components/Composer/Composer.tsx
@@ -133,131 +133,7 @@ const Root = styled(Dialog, {
   name: PREFIX,
   slot: 'Root',
   overridesResolver: (props, styles) => styles.root
-})(({theme}: any) => {
-  let mediaActionBackground = theme.palette.getContrastText(theme.palette.primary.main);
-  if (mediaActionBackground.startsWith('#')) {
-    mediaActionBackground = hexToRgb(mediaActionBackground).replace(')', ', .5)');
-  }
-  return {
-    [`& .${classes.title}`]: {
-      borderBottom: '1px solid #D1D1D1',
-      padding: theme.spacing(),
-      paddingLeft: theme.spacing(2),
-      paddingRight: theme.spacing(2),
-      display: 'flex',
-      flexDirection: 'row',
-      flexWrap: 'nowrap',
-      justifyContent: 'space-between',
-      alignItems: 'center',
-      '& > div': {
-        flex: 1,
-        textAlign: 'center'
-      },
-      '& > div:first-of-type': {
-        textAlign: 'left'
-      },
-      '& > div:last-of-type': {
-        textAlign: 'right',
-        display: 'block'
-      }
-    },
-    [`& .${classes.types}`]: {
-      flexDirection: 'row',
-      alignItems: 'center'
-    },
-    [`& .${classes.avatar}`]: {
-      width: theme.selfcommunity.user.avatar.sizeMedium,
-      height: theme.selfcommunity.user.avatar.sizeMedium,
-      display: 'inline-block'
-    },
-    [`& .${classes.content}`]: {
-      paddingLeft: 0,
-      paddingRight: 0,
-      paddingBottom: 0,
-      position: 'relative',
-      overflowY: 'visible'
-    },
-    [`& .${classes.mediaContent}, & .${classes.audienceContent}, & .${classes.locationContent}`]: {
-      minHeight: 300
-    },
-    [`& .${classes.block}`]: {
-      padding: theme.spacing(2)
-    },
-    [`& .${classes.editor}`]: {
-      minHeight: 200,
-      '& .SCEditor-placeholder': {
-        top: theme.spacing(2),
-        left: theme.spacing(2)
-      }
-    },
-    [`& .${classes.divider}`]: {
-      borderTop: '1px solid #D1D1D1'
-    },
-    [`& .${classes.medias}`]: {
-      margin: '0 23px'
-    },
-    [`& .${classes.location}, & .${classes.audience}`]: {
-      padding: theme.spacing(2),
-      paddingBottom: 0
-    },
-    [`& .${classes.mediasActions}`]: {
-      position: 'absolute',
-      left: 0,
-      right: 0,
-      background: mediaActionBackground,
-      padding: theme.spacing(),
-      zIndex: 1,
-      display: 'flex',
-      flexWrap: 'nowrap',
-      justifyContent: 'space-between',
-      [`& .${classes.mediasActionsActions}`]: {
-        textAlign: 'right'
-      }
-    },
-    [`& .${classes.sortableMedia}`]: {
-      position: 'relative'
-    },
-    [`& .${classes.sortableMediaCover}`]: {
-      backgroundSize: 'cover !important',
-      backgroundPosition: 'center !important',
-      backgroundRepeat: 'no-repeat !important',
-      border: '2px solid white',
-      borderRadius: 6,
-      height: 300
-    },
-    [`& .${classes.links}`]: {
-      padding: theme.spacing(2)
-    },
-    [`& .${classes.actions}`]: {
-      margin: 0,
-      borderTop: '1px solid #D1D1D1',
-      padding: theme.spacing(1),
-      display: 'block',
-      '& .MuiTypography-alignLeft': {
-        float: 'left'
-      },
-      '& .MuiTypography-alignRight': {
-        float: 'right'
-      },
-      [theme.breakpoints.up('md')]: {
-        display: 'flex',
-        flexDirection: 'row',
-        flexWrap: 'nowrap',
-        justifyContent: 'space-between',
-        alignItems: 'center',
-        '& .MuiTypography-alignLeft, & .MuiTypography-alignRight': {
-          float: 'none'
-        }
-      }
-    },
-    [`& .${classes.actionInput}`]: {
-      display: 'none !important'
-    },
-    [`& .${classes.badgeError} .MuiBadge-badge`]: {
-      padding: 0
-    }
-  };
-});
+})(({theme}: any) => ({}));
 
 export interface ComposerTypeMap<P = {}, D extends React.ElementType = 'div'> {
   props: P &
